use new cluster, table -> view
