+++ model/views/sra.py
@@ -2,24 +2,23 @@ from dataclasses import dataclass
 from .. import db
 
 @dataclass
-class analysis_list(db.Model):
+class analysis_index(db.Model):
     run_id : str
-    srarun : str
-    nsra : str
-    psra : str
-    rsra : str
-    assembly : str
-    micro : str
-    geo : str
+    srarun : bool
+    nsra : bool
+    psra : bool
+    rsra : bool
+    assembly : bool
+    micro : bool
+    geo : bool
 
     run_id = db.Column(db.Text, primary_key=True)
-    srarun = db.Column(db.Text)
-    nsra = db.Column(db.Text)
-    psra = db.Column(db.Text)
-    rsra = db.Column(db.Text)
-    assembly = db.Column(db.Text)
-    micro = db.Column(db.Text)
-    geo = db.Column(db.Text)
+    srarun = db.Column(db.Boolean)
+    nsra = db.Column(db.Boolean)
+    psra = db.Column(db.Boolean)
+    rsra = db.Column(db.Boolean)
+    assembly = db.Column(db.Boolean)
+    micro = db.Column(db.Boolean)
+    geo = db.Column(db.Boolean)
 
     filter_col_name = 'run_id'
-
