+++ query/sra.py
@@ -1,5 +1,5 @@
 from .base import QueryBase
-from model.tables.sra import analysis_index
+from model.views.sra import analysis_index
 
 def get_analysis_index(run_id):
     query = analysis_index.query.filter(analysis_index.run_id == run_id)
